chore: optimization, typing big changes
We're starting to work on making our code better. Next, we'll make our functions and classes simpler and more efficient. Sorry for not being active lately!

Our period of inactivity was actually a time for thinking about the concept of Aurum and how to move forward. We'd love to hear your feedback on the library, any suggestions you have, and if you could help with testing.